updted

+++ amafhh/amafhh/doctype/roll_to_sheet_conversion/roll_to_sheet_conversion.js
@@ -108,14 +108,14 @@ frappe.ui.form.on('Roll To Sheet Conversion Items', {
     },
     width_target: function (frm, cdt, cdn) {
         var row = locals[cdt][cdn];
-        if (parseInt(row.width_target) > parseInt(row.width_source)) {
+        if (parseFloat(row.width_target) > parseFloat(row.width_source)) {
             frappe.model.set_value(cdt, cdn, 'width_target', null);
             frappe.throw(__("Target Width cannot be greater than Source Width"));
         } else {
             var conversionType = frm.doc.conversion_type;
             calculateWeightAndSetValues(row, conversionType, cdt, cdn);
         }
-        if (parseInt(row.width_target) < 1) {
+        if (parseFloat(row.width_target) < 1) {
             frappe.model.set_value(cdt, cdn, 'width_target', null);
             frappe.throw(__("Target Width cannot be less than 1"));
         } else {
